+++ src/vue/modules/assets/shared/components/asset-attributes-viewer.vue
@@ -25,7 +25,19 @@
           <tr v-if="balance">
             <td>{{ 'assets.balance-title' | globalize }}</td>
             <td>
-              {{ { value: balance, currency: asset.code } | formatMoney }}
+              <!-- eslint-disable-next-line max-len -->
+              {{ { value: balance.balance, currency: asset.code } | formatMoney }}
+            </td>
+          </tr>
+
+          <tr v-if="balance.isConverted">
+            <td>{{ 'assets.converted-balance-title' | globalize }}</td>
+            <td>
+              {{
+                // eslint-disable-next-line max-len
+                { value: balance.convertedBalance.available, currency: businessStatsQuoteAsset } | formatMoney
+              }}
+              {{ businessStatsQuoteAsset }}
             </td>
           </tr>
 
@@ -67,6 +79,9 @@
 import LogoViewer from './logo-viewer'
 
 import { AssetRecord } from '@/js/records/entities/asset.record'
+import { mapGetters } from 'vuex'
+import { vuexTypes } from '@/vuex'
+import { BalanceRecord } from '@/js/records/entities/balance.record'
 
 const STELLAR_TYPES = {
   creditAlphanum4: 'credit_alphanum4',
@@ -81,11 +96,15 @@ export default {
   },
   props: {
     asset: { type: AssetRecord, required: true },
-    balance: { type: String, default: '' },
+    balance: { type: BalanceRecord, required: true },
     kycRequiredAssetType: { type: Number, required: true },
     securityAssetType: { type: Number, required: true },
   },
   computed: {
+    ...mapGetters({
+      businessStatsQuoteAsset: vuexTypes.businessStatsQuoteAsset,
+    }),
+
     stellarAssetTypeTranslated () {
       let translationId