+++ src/vue/modules/create-asset-form-simplified/mixins/manage-asset-request.mixin.js
@@ -21,7 +21,7 @@ export default {
 
   computed: {
     ...mapGetters([
-      vuexTypes.statsQuoteAsset,
+      vuexTypes.businessStatsQuoteAsset,
       vuexTypes.accountId,
       vuexTypes.accountBalanceByCode,
     ]),
@@ -104,7 +104,7 @@ export default {
     $buildPairCreationRequestOperation () {
       const opts = {
         base: this.collectedAttributes.code,
-        quote: this.statsQuoteAsset.code,
+        quote: this.businessStatsQuoteAsset,
         action: base.xdr.ManageAssetPairAction.create(),
         policies: +ASSET_PAIR_POLICIES.currentPriceRestriction,
         physicalPrice: '' + this.collectedAttributes.price,
@@ -138,7 +138,7 @@ export default {
         quoteAssets: [
           {
             price: this.collectedAttributes.price,
-            asset: this.statsQuoteAsset.code,
+            asset: this.businessStatsQuoteAsset,
           },
         ],
         creatorDetails: {},