+++ src/vue/pages/Security.vue
@@ -160,6 +160,18 @@
         </a>
       </div>
     </template>
+
+    <!-- eslint-disable-next-line max-len -->
+    <template v-if="getModule().canRenderSubmodule(DefaultQuoteAssetPseudoModule)">
+      <div class="security-page__row">
+        <p class="security-page__row-title">
+          {{ 'security-page.default-quote-asset-title' | globalize }}
+        </p>
+        <submodule-importer
+          :submodule="getModule().getSubmodule(DefaultQuoteAssetPseudoModule)"
+        />
+      </div>
+    </template>
   </div>
 </template>
 
@@ -173,6 +185,7 @@ import KeyViewer from '@/vue/common/KeyViewer'
 import ChangePasswordForm from '@/vue/forms/ChangePasswordForm'
 import PhoneNumberForm from '@/vue/forms/PhoneNumberForm'
 import TfaForm from '@/vue/forms/TfaForm'
+import SubmoduleImporter from '@/modules-arch/submodule-importer'
 
 import { ErrorHandler } from '@/js/helpers/error-handler'
 
@@ -185,6 +198,7 @@ import { ShowSeedPseudoModule } from '@/modules-arch/pseudo-modules/show-seed-ps
 import { ChangePasswordPseudoModule } from '@/modules-arch/pseudo-modules/change-password-pseudo-module'
 import { ShowNetworkPassphrasePseudoModule } from '@/modules-arch/pseudo-modules/show-network-passphrase-pseudo-module'
 import { PhoneNumberFormPseudoModule } from '@/modules-arch/pseudo-modules/phone-number-form-pseudo-module'
+import { DefaultQuoteAssetPseudoModule } from '@/modules-arch/pseudo-modules/default-quote-asset-pseudo-module'
 
 const VIEW_MODES = {
   enableTfa: 'enableTfa',
@@ -206,6 +220,7 @@ export default {
     ChangePasswordForm,
     TfaForm,
     PhoneNumberForm,
+    SubmoduleImporter,
   },
   data: _ => ({
     isDrawerShown: false,
@@ -216,6 +231,7 @@ export default {
     ChangePasswordPseudoModule,
     ShowNetworkPassphrasePseudoModule,
     PhoneNumberFormPseudoModule,
+    DefaultQuoteAssetPseudoModule,
     api,
     walletSeed: '',
   }),